api/requirements.txt
Django>=6.0.2
djangorestframework
python-decouple
psycopg2-binary
django-cors-headers
Pillow
django-ckeditor
drf-spectacular
# Optional: for email sending if using a specific backend later
# django-sendgrid-v5
